Add doc comments to explain scrape-examples feature
