+++ src/cargo/core/compiler/job_queue.rs
@@ -348,23 +348,29 @@ enum Message {
     BuildPlanMsg(String, ProcessBuilder, Arc<Vec<OutputFile>>),
     Stdout(String),
     Stderr(String),
+
+    // This is for general stderr output from subprocesses
     Diagnostic {
         id: JobId,
         level: String,
         diag: String,
         fixable: bool,
     },
-    // This is distinct from Diagnostic because it gets emitted through
-    // a different Shell pathway
-    Warning {
-        id: JobId,
-        warning: String,
-    },
+    // This handles duplicate output that is suppressed, for showing
+    // only a count of duplicate messages instead
     WarningCount {
         id: JobId,
         emitted: bool,
         fixable: bool,
     },
+    // This is for warnings generated by Cargo's interpretation of the
+    // subprocess output, e.g. scrape-examples prints a warning if a
+    // unit fails to be scraped
+    Warning {
+        id: JobId,
+        warning: String,
+    },
+
     FixDiagnostic(diagnostic_server::Message),
     Token(io::Result<Acquired>),
     Finish(JobId, Artifact, CargoResult<()>),
@@ -412,6 +418,7 @@ impl<'a, 'cfg> JobState<'a, 'cfg> {
         Ok(())
     }
 
+    /// See [`Message::Diagnostic`] and [`Message::WarningCount`].
     pub fn emit_diag(&self, level: String, diag: String, fixable: bool) -> CargoResult<()> {
         if let Some(dedupe) = self.output {
             let emitted = dedupe.emit_diag(&diag)?;
@@ -433,6 +440,7 @@ impl<'a, 'cfg> JobState<'a, 'cfg> {
         Ok(())
     }
 
+    /// See [`Message::Warning`].
     pub fn warning(&self, warning: String) -> CargoResult<()> {
         self.messages.push_bounded(Message::Warning {
             id: self.id,
