+++ src/cargo/core/compiler/rustdoc.rs
@@ -191,6 +191,8 @@ pub fn add_root_urls(
     Ok(())
 }
 
+/// Indicates whether a target should have examples scraped from it
+/// by rustdoc. Configured within Cargo.toml.
 #[derive(Clone, Hash, PartialEq, Eq, PartialOrd, Ord, Debug, Copy)]
 pub enum RustdocScrapeExamples {
     Enabled,
